+++ examples/axis_inputs.rs
@@ -62,7 +62,7 @@ fn move_player(query: Query<&ActionState<Action>, With<Player>>) {
         //
         // If you don't have a variable trigger, this will just return 0.0 when not pressed and 1.0
         // when pressed.
-        let value = action_state.clamped_value(&Action::Throttle);
+        let value = action_state.clamped_button_value(&Action::Throttle);
         println!("Throttle: {value}");
     }